Add logic/Math for calculating distance and longest.

+++ golfTracker/Data Manager/Clubs.swift
@@ -10,7 +10,7 @@ import RealmSwift
 
 class Clubs: Object{
     @objc dynamic var name: String = ""
-    @objc dynamic var avgDistance: Float = 0.0
+    @objc dynamic var avgDistance: Double = 0.0
     @objc dynamic var longestDistance: Int = 0
     @objc dynamic var amountHit: Int = 0
     
@@ -20,7 +20,7 @@ class Clubs: Object{
 class Putter: Object{
     @objc dynamic var name: String = ""
     @objc dynamic var amountHit: Int = 0
-    @objc dynamic var avgPutts: Float = 0.0
+    @objc dynamic var avgPutts: Double = 0.0
 }
