+++ golfTracker/Data Manager/Scores.swift
@@ -10,7 +10,7 @@ import RealmSwift
 
 class Scores: Object{
     @objc dynamic var lowestScore: Int = 0
-    @objc dynamic var avgScore: Float = 0.0
+    @objc dynamic var avgScore: Double = 0.0
     @objc dynamic var totalRounds: Int = 0
     
     
@@ -20,7 +20,7 @@ class RoundStats: Object{
     
     @objc dynamic var greenInReg: Int = 0
     @objc dynamic var fairwaysInReg: Int = 0
-    @objc dynamic var numberOfPutts: Float = 0
-    @objc dynamic var girAvg: Float = 0.0
-    @objc dynamic var firAvg: Float = 0.0
+    @objc dynamic var numberOfPutts: Double = 0
+    @objc dynamic var girAvg: Double = 0.0
+    @objc dynamic var firAvg: Double = 0.0
 }
